platforms/common: Move DRMEventHandler in from eglstream-kms.
This should be useful when transitioning to the Atomic KMS API in gbm-kms.
