+++ src/platforms/common/server/kms-utils/CMakeLists.txt
@@ -6,6 +6,9 @@ add_library(${KMS_UTILS_STATIC_LIBRARY} STATIC
   drm_mode_resources.h
   kms_connector.cpp
   kms_connector.h
+  drm_event_handler.h
+  threaded_drm_event_handler.h
+  threaded_drm_event_handler.cpp
 )
 
 target_include_directories(${KMS_UTILS_STATIC_LIBRARY}
@@ -14,4 +17,5 @@ target_include_directories(${KMS_UTILS_STATIC_LIBRARY}
 
 target_link_libraries(${KMS_UTILS_STATIC_LIBRARY}
   PkgConfig::DRM
+  mircore
 )
