+++ tests/unit-tests/platforms/eglstream-kms/server/CMakeLists.txt
@@ -1,6 +1,5 @@
 list(APPEND EGLSTREAM_KMS_UNIT_TEST_SOURCES
   $<TARGET_OBJECTS:mirplatformgraphicseglstreamkmsobjects>
-  ${CMAKE_CURRENT_SOURCE_DIR}/test_threaded_drm_event_handler.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/test_utils.cpp
 )
 set(EGLSTREAM_KMS_UNIT_TEST_SOURCES ${EGLSTREAM_KMS_UNIT_TEST_SOURCES} PARENT_SCOPE)
